requirements.txt
click==8.1.7
pandas==2.1.3
python-dotenv==1.0.0
tqdm==4.66.1
requests==2.31.0
xmltodict==0.13.0
datetime==5.2
Pillow==10.1.0
bs4==0.0.1
openpyxl==3.1.2
